testing/requirements.txt
pytest
requests
python-dotenv
flake8
black
pydantic>=2.0.0
